V1.4 Change Tb_user's total_storage and used_storage's type from float to double.

+++ src/com/bjtu/ajax/file/DeleteListInGalleryAction.java
@@ -40,7 +40,7 @@ public class DeleteListInGalleryAction extends ActionSupport{
 		Tb_file file;
 		FileSystem fs=HDFSUtil.openFileSystem();
 		//初始化被删除文件总大小
-		float total_delete_size=0;
+		double total_delete_size=0;
 		while(st.hasMoreElements()){
 			id=st.nextToken();
 			file=file_service.getById(id);
